# apps/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ValidationErrorDetail"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "TaskCategory": {
        "enum": [
          "work",
          "personal",
          "study",
          "other"
        ],
        "title": "TaskCategory",
        "type": "string"
      },
      "TaskCreate": {
        "additionalProperties": false,
        "description": "`POST /api/v1/tasks` request body (spec §6.1 defaults).\n\nExtra fields are rejected outright — both truly unknown keys and every\nserver-owned field (`id`, `position`, `created_at`, `updated_at`,\n`completed_at`, `archived_at`), none of which this model declares, so\nthey can never be silently dropped or, worse, silently honored.",
        "properties": {
          "category": {
            "$ref": "#/components/schemas/TaskCategory",
            "default": "other"
          },
          "content": {
            "title": "Content",
            "type": "string"
          },
          "deadline_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deadline At"
          },
          "markdown_note": {
            "default": "",
            "title": "Markdown Note",
            "type": "string"
          },
          "priority": {
            "$ref": "#/components/schemas/TaskPriority",
            "default": "medium"
          },
          "status": {
            "$ref": "#/components/schemas/TaskStatus",
            "default": "todo"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "urls": {
            "items": {
              "$ref": "#/components/schemas/TaskUrl"
            },
            "title": "Urls",
            "type": "array"
          }
        },
        "required": [
          "title",
          "content"
        ],
        "title": "TaskCreate",
        "type": "object"
      },
      "TaskMove": {
        "additionalProperties": false,
        "description": "`POST /api/v1/tasks/{task_id}/move` request body (issue #29).\n\n`index` is an insertion index, not a stored `position` value — for a\nmove into another column it is the index among that column's existing\ntasks (`0` to `len(target)` inclusive); for a move whose `status`\nequals the task's current one it is the task's final index in its own\ncolumn (`0` to `len(column) - 1` inclusive). Both bounds depend on the\ntarget column's size at request time, so only the lower bound (never\nnegative) is enforced by the wire type here — the upper bound is\nchecked by `domain/ordering.py` and turned into a `422` by the router;\nneither bound is ever clamped.",
        "properties": {
          "index": {
            "minimum": 0.0,
            "title": "Index",
            "type": "integer"
          },
          "status": {
            "$ref": "#/components/schemas/TaskStatus"
          }
        },
        "required": [
          "status",
          "index"
        ],
        "title": "TaskMove",
        "type": "object"
      },
      "TaskPriority": {
        "enum": [
          "low",
          "medium",
          "high"
        ],
        "title": "TaskPriority",
        "type": "string"
      },
      "TaskReorder": {
        "additionalProperties": false,
        "description": "`POST /api/v1/tasks/reorder` request body (issue #29).\n\n`ordered_ids` must be an exact permutation of the ids of the active\ntasks currently in `status` — checked by `domain/ordering.reorder_column`\nand turned into a `422` by the router for a duplicate, an omission, or\na foreign/archived/unknown id.",
        "properties": {
          "ordered_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Ordered Ids",
            "type": "array"
          },
          "status": {
            "$ref": "#/components/schemas/TaskStatus"
          }
        },
        "required": [
          "status",
          "ordered_ids"
        ],
        "title": "TaskReorder",
        "type": "object"
      },
      "TaskResponse": {
        "additionalProperties": false,
        "description": "The full task wire shape — exactly the 14 §5 fields, nothing else.",
        "properties": {
          "archived_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archived At"
          },
          "category": {
            "$ref": "#/components/schemas/TaskCategory"
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "content": {
            "title": "Content",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "deadline_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deadline At"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "markdown_note": {
            "title": "Markdown Note",
            "type": "string"
          },
          "position": {
            "title": "Position",
            "type": "number"
          },
          "priority": {
            "$ref": "#/components/schemas/TaskPriority"
          },
          "status": {
            "$ref": "#/components/schemas/TaskStatus"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "urls": {
            "items": {
              "$ref": "#/components/schemas/TaskUrl"
            },
            "title": "Urls",
            "type": "array"
          }
        },
        "required": [
          "id",
          "title",
          "content",
          "status",
          "category",
          "priority",
          "deadline_at",
          "urls",
          "markdown_note",
          "position",
          "created_at",
          "updated_at",
          "completed_at",
          "archived_at"
        ],
        "title": "TaskResponse",
        "type": "object"
      },
      "TaskStatus": {
        "enum": [
          "todo",
          "in_progress",
          "done"
        ],
        "title": "TaskStatus",
        "type": "string"
      },
      "TaskUpdate": {
        "additionalProperties": false,
        "description": "`PATCH /api/v1/tasks/{id}` request body — a partial update.\n\nEvery field defaults to `None`, meaning \"left out of the body\" — the\nrouter distinguishes that from an *explicit* `null` via\n`model_dump(exclude_unset=True)`, since pydantic does not run a field's\nvalidators against a default value that was never supplied (only\nagainst a value actually present in the request, including an explicit\n`null`). An explicit `null` for a field the database stores `NOT NULL`\n(`title`, `content`, `markdown_note`, `category`, `priority`, `status`,\n`urls`) is rejected by the validators below; only `deadline_at` accepts\nan explicit `null`, which clears the deadline.",
        "properties": {
          "category": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TaskCategory"
              },
              {
                "type": "null"
              }
            ]
          },
          "content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content"
          },
          "deadline_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deadline At"
          },
          "markdown_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Markdown Note"
          },
          "priority": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TaskPriority"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TaskStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "urls": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/TaskUrl"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Urls"
          }
        },
        "title": "TaskUpdate",
        "type": "object"
      },
      "TaskUrl": {
        "additionalProperties": false,
        "description": "One URL entry (spec §5, §8). No `id` on the wire — the web tier's\nHTTP mapper assigns a client-side id for its own `TaskUrl.id` (the\nuser's non-scope decision recorded on issue #28: URL entries stay\nid-less on the wire).",
        "properties": {
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "url"
        ],
        "title": "TaskUrl",
        "type": "object"
      },
      "ValidationErrorDetail": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "field",
          "code",
          "message"
        ],
        "title": "ValidationErrorDetail",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Planora API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/tasks": {
      "get": {
        "operationId": "list_tasks_api_v1_tasks_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TaskResponse"
                  },
                  "title": "Response List Tasks Api V1 Tasks Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "API error envelope"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request fields"
          }
        },
        "summary": "List Tasks",
        "tags": [
          "tasks"
        ]
      },
      "post": {
        "operationId": "create_task_api_v1_tasks_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "API error envelope"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "API error envelope"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request fields"
          }
        },
        "summary": "Create Task",
        "tags": [
          "tasks"
        ]
      }
    },
    "/api/v1/tasks/reorder": {
      "post": {
        "operationId": "reorder_tasks_api_v1_tasks_reorder_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskReorder"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TaskResponse"
                  },
                  "title": "Response Reorder Tasks Api V1 Tasks Reorder Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "API error envelope"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "API error envelope"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request fields"
          }
        },
        "summary": "Reorder Tasks",
        "tags": [
          "tasks"
        ]
      }
    },
    "/api/v1/tasks/{task_id}": {
      "delete": {
        "operationId": "delete_task_api_v1_tasks__task_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "API error envelope"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "API error envelope"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "API error envelope"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request fields"
          }
        },
        "summary": "Delete Task",
        "tags": [
          "tasks"
        ]
      },
      "get": {
        "operationId": "get_task_api_v1_tasks__task_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "API error envelope"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "API error envelope"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request fields"
          }
        },
        "summary": "Get Task",
        "tags": [
          "tasks"
        ]
      },
      "patch": {
        "operationId": "update_task_api_v1_tasks__task_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "API error envelope"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "API error envelope"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "API error envelope"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request fields"
          }
        },
        "summary": "Update Task",
        "tags": [
          "tasks"
        ]
      }
    },
    "/api/v1/tasks/{task_id}/move": {
      "post": {
        "description": "Board drag (or an AI-confirmed move), cross-column or within one\ncolumn (issue #29). `completed_at` always goes through\n`resolve_completed_at` — entering, leaving and staying in Done are all\nhandled by that one rule, never re-implemented here.",
        "operationId": "move_task_api_v1_tasks__task_id__move_post",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskMove"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TaskResponse"
                  },
                  "title": "Response Move Task Api V1 Tasks  Task Id  Move Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "API error envelope"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "API error envelope"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "API error envelope"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request fields"
          }
        },
        "summary": "Move Task",
        "tags": [
          "tasks"
        ]
      }
    }
  }
}